+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Upcoming
+* Changed Quaron bonus speed on wrong blocks (+1 -> +0.5 + 0.2/efficiency level)
 * Mobs from spawners now have reduced buffs
 * Wolves and Trader llamas are now 10% slower